Conflicts:
	edexOsgi/build.edex/esb/data/utility/edex_static/base/python/ObsRequest.py
	edexOsgi/com.raytheon.uf.common.dataplugin/src/com/raytheon/uf/common/dataplugin/persist/DefaultPathProvider.java
	edexOsgi/com.raytheon.uf.common.dataplugin/src/com/raytheon/uf/common/dataplugin/persist/PersistablePluginDataObject.java

Former-commit-id: 049bdc378d34047fe9c57752b2c25b947611e7fc [formerly 6a1f2160bde22da06358a1263047a67b19f028fc] [formerly c90ce9d54eb7febd9e7ed5e60ef3f8baceb524f1] [formerly c90ce9d54eb7febd9e7ed5e60ef3f8baceb524f1 [formerly ed26e06df222d4df26176eb6fdaa52ae5900e64a]] [formerly fee53dea94d79cd2af4d3c9146719eb67bebaa2b [formerly c90ce9d54eb7febd9e7ed5e60ef3f8baceb524f1 [formerly ed26e06df222d4df26176eb6fdaa52ae5900e64a] [formerly fee53dea94d79cd2af4d3c9146719eb67bebaa2b [formerly 3bd9b3a16b0105822983302efa5dd2834aaf5a5f]]]]
Former-commit-id: fee53dea94d79cd2af4d3c9146719eb67bebaa2b
Former-commit-id: 6be16a9e5ed9f90b18ddcd7a91a8e02eafd51c8e [formerly 530c02831155693cd7d1ae3b27691f3302b2bbe3] [formerly 51bb9ebb4672c043d3b9845710f43426ffb6b744 [formerly dcdd7f0648c96a3fd4e50b2d74706897c71b80e5]]
Former-commit-id: 51bb9ebb4672c043d3b9845710f43426ffb6b744
Former-commit-id: a4eb467a3411a75f5f70fb2b27f597fcb381c65c

+++ rpms/awips2.core/Installer.notification/component.spec
@@ -39,8 +39,9 @@ fi
 mkdir -p ${RPM_BUILD_ROOT}/awips2/notification
 mkdir -p ${RPM_BUILD_ROOT}/etc/profile.d
 
-PROFILE_D_DIR="Installer.rpm/awips2.core/Installer.notification/scripts/profile.d"
-cp ${WORKSPACE_DIR}/${PROFILE_D_DIR}/* ${RPM_BUILD_ROOT}/etc/profile.d
+RPM_CORE_PROJECT_DIR="${WORKSPACE_DIR}/Installer.rpm/awips2.core"
+PROFILE_D_DIR="${RPM_CORE_PROJECT_DIR}/Installer.notification/scripts/profile.d"
+cp ${PROFILE_D_DIR}/* ${RPM_BUILD_ROOT}/etc/profile.d
 
 %build
 #---------------------------------------------------------------------------#
@@ -76,13 +77,21 @@ function copyLegal()
       
    rm -f ${WORKSPACE_DIR}/Installer.rpm/legal/FOSS_licenses.tar    
 }
-NOTIFICATION_TAR_FILE_DIR="packages/notification"
+RPM_CORE_PROJECT_DIR="${WORKSPACE_DIR}/Installer.rpm/awips2.core"
+NOTIFICATION_TAR_FILE_DIR="${RPM_CORE_PROJECT_DIR}/Installer.notification/src"
 NOTIFICATION_TAR_FILE="${NOTIFICATION_TAR_FILE_DIR}/edex_com.tar.bz2"
 
 cd ${RPM_BUILD_ROOT}/awips2
-/bin/gtar -xpf ${AWIPSCM_SHARE}/${NOTIFICATION_TAR_FILE}
+/bin/gtar -xpf ${NOTIFICATION_TAR_FILE}
+if [ $? -ne 0 ]; then
+   exit 1
+fi
 
-cp -r ${RPM_BUILD_ROOT}/awips2/edex_com/* ${RPM_BUILD_ROOT}/awips2/notification/
+cp -r ${RPM_BUILD_ROOT}/awips2/edex_com/* \
+   ${RPM_BUILD_ROOT}/awips2/notification/
+if [ $? -ne 0 ]; then
+   exit 1
+fi
 # Remove the boost rpms directory
 rm -rf ${RPM_BUILD_ROOT}/awips2/notification/rpms
 rm -rf ${RPM_BUILD_ROOT}/awips2/edex_com
